requirements.txt
CherryPy==4.0.0
Django==1.8.4
Fabric==1.10.2
Pillow==3.0.0
PyYAML==3.11
Pygments==2.0.2
WebOb==1.4.1
WebTest==2.0.18
Werkzeug==0.10.4
appnope==0.1.0
beautifulsoup4==4.4.1
blessings==1.6
bpython==0.14.2
contextlib2==0.4.0
curtsies==0.1.19
decorator==4.0.4
dj-database-url==0.3.0
dj-static==0.0.6
django-address==0.1.3
django-allauth==0.23.0
django-common==0.1.51
django-extensions==1.5.7
django-image-tools-2==1.0
django-localflavor==1.2
django-queryset-csv==0.3.3
django-tinymce==2.2.0
django-toolbelt==0.0.1
django-vanilla-views==1.0.4
django-webtest==1.7.8
django-widget-tweaks==1.4.1
django-wysiwyg==0.7.1
djrill==2.0
ecdsa==0.13
gnureadline==6.3.3
greenlet==0.4.9
gunicorn==19.3.0
httpretty==0.8.6
ipdb==0.8.1
ipython-genutils==0.1.0
ipython==4.0.0
mock==1.0.0
model-mommy==1.2.5
newrelic==2.60.0.46
oauthlib==1.0.3
paramiko==1.16.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
psycopg2==2.6.1
pycrypto==2.6.1
python-dateutil==2.4.2
python-openid==2.2.5
raven==5.9.2
requests-oauthlib==0.5.0
requests==2.7.0
silk-config==0.3.5
silk-deployment==0.3.14
simplegeneric==0.8.1
six==1.9.0
static3==0.6.1
traitlets==4.0.0
waitress==0.8.10
wsgiref==0.1.2
xmltodict==0.9.2
